requirements.txt
asgiref==3.2.7
astroid==2.4.2
boto3==1.14.40
botocore==1.17.40
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-filter==2.3.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.7.0
isort==4.3.21
jmespath==0.10.0
jsonpickle==1.4.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.1.2
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.1.0
